refactor: Улучшение организации кода - Разделение кода на модули для улучшения читаемости и повторного использования. - Перенос функций детекции в detection.py. - Перенос функций отправки сообщений в telegram_utils.py. - Основной скрипт теперь main.py.
Разделение кода на модули улучшает его структуру и делает проект более поддерживаемым и расширяемым.
